created topic-pool component and fixed subtopic loading
