+++ src/app/models/curriculum-day.ts
@@ -2,6 +2,9 @@ import { SubTopic } from './subtopic';
 export class CurriculumDay {
     curriculumDayId?: number;
     curriculumWeekId?: number;
+    /**
+     * 0 - 6, Sunday - Saturday
+     */
     dayNum: number;
     supTopics: SubTopic[];
 }
