+++ src/app/models/curriculum.ts
@@ -7,6 +7,10 @@ export class Curriculum {
     version: number;
     creatorId: number;
     approvedById?: number;
+    /**
+     * 1 - draft, 2 - needs approval, 3 - read-only, 4 - master
+     */
+    status: number;
     dateCreated: Date;
     numberOfWeeks: number;
     topics: Topic[];
