+++ src/app/routes/curriculum-editor/curriculum-day/curriculum-day.component.html
@@ -1,29 +1,29 @@
-<mat-card *ngIf="day">
+<div class="column" *ngIf="day">
   <mat-card-title-group>
-    <mat-card-title *ngIf="day.dayNum == 1">
+    <mat-card-title *ngIf="day.dayNum == 0">
       Sunday
     </mat-card-title>
-    <mat-card-title *ngIf="day.dayNum == 2">
+    <mat-card-title *ngIf="day.dayNum == 1">
       Monday
     </mat-card-title>
-    <mat-card-title *ngIf="day.dayNum == 3">
+    <mat-card-title *ngIf="day.dayNum == 2">
       Tuesday
     </mat-card-title>
-    <mat-card-title *ngIf="day.dayNum == 4">
+    <mat-card-title *ngIf="day.dayNum == 3">
       Wednesday
     </mat-card-title>
-    <mat-card-title *ngIf="day.dayNum == 5">
+    <mat-card-title *ngIf="day.dayNum == 4">
       Thursday
     </mat-card-title>
-    <mat-card-title *ngIf="day.dayNum == 6">
+    <mat-card-title *ngIf="day.dayNum == 5">
       Friday
     </mat-card-title>
-    <mat-card-title *ngIf="day.dayNum == 7">
+    <mat-card-title *ngIf="day.dayNum == 6">
       Saturday
     </mat-card-title>
-    <mat-card-title *ngIf="day.dayNum > 7 && day.dayNum < 1">
+    <mat-card-title *ngIf="day.dayNum > 6 && day.dayNum < 0">
       Unknown Day Num: {{day.dayNum}}
     </mat-card-title>
   </mat-card-title-group>
   Subtopics go here
-</mat-card>
+</div>
