+++ src/app/routes/curriculum-editor/curriculum-editor.component.html
@@ -57,126 +57,7 @@
           </mat-card>
         </mat-tab>
         <mat-tab label="Topic Pool">
-          <mat-card>
-            <mat-card-header>
-              <mat-toolbar class="panel-bar">
-                <span style="flex: 0.3 1 auto"></span>
-                <!-- <h4 class="m-0 p-0" style="font-size: 140%;">Topic Pool</h4> -->
-                <span style="flex: 1 1 auto"></span>
-                <mat-form-field class="my-0 pb-0" style="color: grey;">
-                  <input style="color: white;" matInput color="accent" type="text" placeholder="Search Subtopic"
-                    [(ngModel)]="topicSearch">
-                  <button mat-button *ngIf="value" matSuffix mat-icon-button aria-label="Clear" (click)="value=''">
-                    <mat-icon>close</mat-icon>
-                  </button>
-                </mat-form-field>
-                <span>
-                  <button mat-button color="accent" mat-icon-button aria-label="add" (click)="openNewDialog()">
-                    <mat-icon>add</mat-icon>
-                  </button>
-                </span>
-                <span style="flex: 0.3 1 auto"></span>
-              </mat-toolbar>
-            </mat-card-header>
-            <mat-card-content style="overflow: auto; height: 470px;">
-              <mat-accordion *ngIf="topicSearch.length < 2">
-                <mat-card-content>
-                  <div *ngFor="let topic of topics">
-                    <mat-expansion-panel [disabled]="topic.name !== topicService.reactivateName(topic.name)" [expanded]="topic.name === topicService.reactivateName(topic.name) &amp;&amp; isExpanded(topic.id)"
-                      (opened)="expandTopic(topic.id)" (closed)="collapseTopic(topic.id)">
-                      <mat-expansion-panel-header>
-                        <mat-panel-title *ngIf="topic.name === topicService.reactivateName(topic.name)">
-
-                          <span>
-                            {{topic.name}}
-                          </span>
-                          <span class="fill-remaining-space">
-                          </span>
-                          <span>
-                            <button mat-icon-button color="primary" (click)="deactivateTopic(topic)">
-                              <mat-icon color="primary">clear</mat-icon>
-                            </button>
-                          </span>
-
-                        </mat-panel-title>
-                        <mat-panel-title *ngIf="topic.name !== topicService.reactivateName(topic.name)">
-                          <span>
-                            <del>{{topicService.reactivateName(topic.name)}}</del>
-                          </span>
-                          <span class="fill-remaining-space">
-                          </span>
-                          <span>
-                            <button mat-icon-button color="" (click)="reactivateTopic(topic)">
-                              <mat-icon color="">add</mat-icon>
-                            </button>
-                          </span>
-                        </mat-panel-title>
-                      </mat-expansion-panel-header>
-                      <div *ngFor="let subtopic of getSubtopicsByTopic(topic)">
-                        <div *ngIf="(!topicSearch || ((subtopic.name | lowercase).indexOf(topicSearch | lowercase) >= 0)) &amp;&amp; (subtopicService.reactivateName(subtopic.name) === subtopic.name)">
-                          <mat-toolbar>
-                            <span> {{(subtopic.name | slice:0:50) + ((subtopic.name.length > 50) ? "..." : "")}}</span>
-                            <span class="fill-remaining-space"></span>
-                            <span>
-                              <button mat-icon-button color="" (click)="deactivateSubtopic(subtopic)">
-                                <mat-icon color="primary">clear</mat-icon>
-                              </button>
-                            </span>
-                          </mat-toolbar>
-                        </div>
-                        <div *ngIf="(!topicSearch || ((subtopic.name | lowercase).indexOf(topicSearch | lowercase) >= 0)) &amp;&amp; (subtopicService.reactivateName(subtopic.name) !== subtopic.name)">
-                          <mat-toolbar>
-                            <span>
-                              <del>{{ subtopicService.reactivateName(subtopic.name) }}</del>
-                            </span>
-                            <span class="fill-remaining-space"> </span>
-                            <span>
-                              <button mat-icon-button color="" (click)="reactivateSubtopic(subtopic)">
-                                <mat-icon color="">add</mat-icon>
-                              </button>
-                            </span>
-                          </mat-toolbar>
-                        </div>
-                      </div>
-                      <div *ngIf="getSubtopicsByTopic(topic).length == 0">
-                        No Subtopics for {{ topic.name }}
-                      </div>
-                    </mat-expansion-panel>
-                  </div>
-                </mat-card-content>
-              </mat-accordion>
-              <mat-accordion *ngIf="topicSearch.length > 1">
-                <mat-card-content>
-                  <div *ngFor="let topic of searchTopics(topicSearch)">
-                    <span>
-                      {{topic.name}}
-                    </span>
-                    <span class="fill-remaining-space">
-                    </span>
-                    <span>
-                      <button mat-icon-button color="primary" (click)="deactivateTopic(topic)">
-                        <mat-icon color="primary">clear</mat-icon>
-                      </button>
-                    </span>
-                    <mat-list *ngIf="hasTopic(topicSearch, topic)">
-                      <mat-list-item *ngFor="let subtopic of searchSubtopics(topicSearch)">
-                        <span>
-                          {{subtopic.name}}
-                        </span>
-                        <span class="fill-remaining-space">
-                        </span>
-                        <span>
-                          <button mat-icon-button color="primary" (click)="deactivateTopic(topic)">
-                            <mat-icon color="primary">clear</mat-icon>
-                          </button>
-                        </span>
-                      </mat-list-item>
-                    </mat-list>
-                  </div>
-                </mat-card-content>
-              </mat-accordion>
-            </mat-card-content>
-          </mat-card>
+          <app-topic-pool></app-topic-pool>
         </mat-tab>
       </mat-tab-group>
