+++ src/app/routes/curriculum-editor/topic-pool/topic-pool.component.spec.ts
@@ -1,0 +1,25 @@
+import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+
+import { TopicPoolComponent } from './topic-pool.component';
+
+describe('TopicPoolComponent', () => {
+  let component: TopicPoolComponent;
+  let fixture: ComponentFixture<TopicPoolComponent>;
+
+  beforeEach(async(() => {
+    TestBed.configureTestingModule({
+      declarations: [ TopicPoolComponent ]
+    })
+    .compileComponents();
+  }));
+
+  beforeEach(() => {
+    fixture = TestBed.createComponent(TopicPoolComponent);
+    component = fixture.componentInstance;
+    fixture.detectChanges();
+  });
+
+  it('should create', () => {
+    expect(component).toBeTruthy();
+  });
+});
